modify matched slot to account for upcoming sessions, WIP
