+++ src/App.js
@@ -1,16 +1,17 @@
 // import axios from './fakeAxios';
-import {Typography, Button, Box} from '@material-ui/core';
+import {Typography, Box} from '@material-ui/core';
 import axios from 'axios';
 import moment from 'moment';
 import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import MainRouter from './components/MainRouter';
-import { ThemeProvider, createMuiTheme, withStyles } from '@material-ui/core/styles';
+import { ThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 import useApplicationData from 'hooks/useApplicationData';
 import ContextContainer from 'contexts/ContextContainer';
+import { fifteenMinutesInMs } from "helpers/constants";
+import JoinButton from "components/Buttons/JoinButton";
 
 const BASE_URL = process.env.REACT_APP_BASE_URL;
-const fifteenMinutesInMs = 1000 * 60 * 15;
 
 const theme = createMuiTheme({
   palette: {
@@ -58,23 +59,6 @@ function App() {
   const { user, setUser, cookies, setCookie, removeCookie } = useApplicationData();
   const history = useHistory();
 
-  const JoinButton = withStyles({
-    root: {
-      backgroundColor: "#23262A",
-      color: "#FFFFFF",
-      '&:hover': {
-        backgroundColor: "black",
-        color: "#FFFFFF",
-        boxShadow: 'none',
-      },
-      '&:active': {
-        boxShadow: 'none',
-        color: "#FFFFFF",
-        backgroundColor: "grey",
-      },
-    }
-  })(Button)
-
   // Renders the upcoming sessions displayed in the sidenav - need to put in App.js so we can include it in the global object and have it be called from calendar actions
   const renderUpcoming = function() {
     if (!user) return;
@@ -86,7 +70,6 @@ function App() {
             const otherUser = users.find( user_ => user_.user_id !== user.user_id);
             const startTimeObj = moment(session.start_time);
             const msToStart = Math.abs(moment().diff(startTimeObj));
-            console.log(msToStart);
             return (
               <>
                 <Box display="flex" justifyContent="space-between">
