+++ src/helpers/constants.js
@@ -1,0 +1,1 @@
+export const fifteenMinutesInMs = 1000 * 60 * 15;
